Move all reddit media commands from main bot script to a cog
